Convert tree-select to typescript.
